requirements.txt
asgiref==3.2.3
Brotli==1.0.7
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.3
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jsonconfig-tool==1.5
keyring==21.1.0
MarkupSafe==1.1.1
messages==0.5.0
openapi-codec==1.3.2
packaging==20.1
psycopg2-binary==2.8.4
pyparsing==2.4.6
python-box==4.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
validus==0.3.0
whitenoise==5.0.1
